Cria mais testes unitarios e adiciona imagem de logo
Cria testes unitarios para a controller, refatora testes do model, inicia testes para o viewmodel, refatora tratamento de erro de chamadas da api.

+++ Challenge-ios/Challenge-ios/View/Controller/HomeTableViewDataSource+Extension.swift
@@ -65,9 +65,9 @@ extension HomeViewController:  UITableViewDataSource {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         switch indexPath.section {
             case 3:
-                return CGFloat(Constants.cellPackageHeight)
+                return Constants.cellPackageHeight
             default:
-                return CGFloat(Constants.cellHotelHeight)
+                return Constants.cellHotelHeight
         }
     }
